+++ bpwave/signal.py
@@ -21,6 +21,9 @@ class CpIndices:
     """Indices of the characteristic points of a single period of an ABP signal.
 
     Fields are optional; missing entries get negative value.
+
+    .. warning::
+        Please don't use negative (backwards) indexing, set values must be >= 0.
     """
 
     NAMES: _t.ClassVar[tuple[str, str, str, str, str, str]] = (
@@ -33,59 +36,81 @@ class CpIndices:
     )
     """Index names in the same order as values in :meth:`to_array`."""
 
-    onset: int = -1
+    UNSET: _t.ClassVar[int] = -1
+    """Value representing unset value.
+
+    .. versionadded:: 0.0.3
+    """
+
+    onset: int = UNSET
     """Onset."""
 
-    sys_peak: int = -1
+    sys_peak: int = UNSET
     """Systolic peak."""
 
-    refl_onset: int = -1
+    refl_onset: int = UNSET
     """Onset of the reflected wave."""
 
-    refl_peak: int = -1
+    refl_peak: int = UNSET
     """Peak of the reflected wave."""
 
-    dicr_notch: int = -1
+    dicr_notch: int = UNSET
     """Dicrotic notch."""
 
-    dicr_peak: int = -1
+    dicr_peak: int = UNSET
     """Peak of the dicrotic wave."""
 
     def __post_init__(self):
-        if (self.to_array() < 0).all():
+        if (self.to_array() == self.UNSET).all():
             warnings.warn("None of the indices are set.")
 
-    def __sub__(self, shift: int) -> "CpIndices":
-        """Subtracts the same scalar from all indices."""
+    def __add__(self, shift: int) -> "CpIndices":
+        """Adds the same scalar to all indices.
+
+        A possible use case is when characteristic points are detected on a
+        shorter section of a signal, but the result needs to be trasferred to
+        the full length one.
+
+        .. versionadded:: 0.0.3
+        """
         return CpIndices(
             **{
-                n: shifted if (shifted := v - shift) >= 0 else -1
+                n: shifted if (shifted := v + shift) >= 0 else self.UNSET
                 for n, v in self.without_unset().items()
             }
         )
 
+    def __sub__(self, shift: int) -> "CpIndices":
+        """Subtracts the same scalar from all indices.
+
+        .. versionadded:: 0.0.3
+        """
+        return self + -shift
+
     def min(self) -> int:
         """Smallest set index."""
-        return min(self.without_unset().values(), default=-1)
+        return min(self.without_unset().values(), default=self.UNSET)
 
     def max(self) -> int:
         """Largest set index."""
-        return max(self.without_unset().values(), default=-1)
+        return max(self.without_unset().values(), default=self.UNSET)
 
     def clamped(self, start: int | None = None, stop: int | None = None) -> "CpIndices":
         """Keeps only the indices contained in the specified range."""
         start = start or 0
-        stop = stop or _np.iinfo(int).max
+        stop = _np.iinfo(int).max if stop is None else stop
         return CpIndices(
             **{
-                n: v if start <= v < stop else -1
+                n: v if start <= v < stop else self.UNSET
                 for n, v in self.without_unset().items()
             }
         )
 
     def without_unset(self) -> dict[str, int]:
         """Returns the existing points in a ``dict``."""
-        return {name: value for name, value in self.__dict__.items() if value >= 0}
+        return {
+            name: value for name, value in self.__dict__.items() if value != self.UNSET
+        }
 
     def to_array(self) -> _npt.NDArray[_np.int64]:
         """The points as numpy array."""
@@ -107,6 +132,97 @@ class ChPoints:
 
     params: dict[str, _t.Any]
     """Parameters of the algorithm, if any."""
+
+    def __getitem__(self, slc: slice) -> "ChPoints":
+        """Returns a copy with sliced ``indices``.
+
+        Negative (backwards) indices and setting ``step`` is not supported.
+
+        .. versionadded:: 0.0.3
+        """
+        if not isinstance(slc, slice):
+            raise NotImplementedError("Only slicing is supported")
+        if slc.step is not None:
+            raise NotImplementedError("Step is not yet supported")
+        if (slc.start is not None and slc.start < 0) or (
+            slc.stop is not None and slc.stop < 0
+        ):
+            raise NotImplementedError("Negative indices not supported")
+
+        start = slc.start or 0
+        stop = _np.iinfo(int).max if slc.stop is None else slc.stop
+
+        return _dc.replace(
+            self,
+            indices=[
+                ci.clamped(start, stop) - start
+                for ci in self.indices
+                if (
+                    (ci.min() >= start and ci.max() < stop)
+                    or (ci.min() <= start <= ci.max())
+                    or (ci.min() < stop <= ci.max())
+                )
+            ],
+        )
+
+    def __add__(self, shift: int) -> "ChPoints":
+        """Adds the same scalar to all indices.
+
+        A possible use case is when characteristic points are detected on a
+        shorter section of a signal, but the result needs to be trasferred to
+        the full length one.
+
+        .. versionadded:: 0.0.3
+        """
+        return _dc.replace(
+            self,
+            indices=[ci + shift for ci in self.indices],
+        )
+
+    def __sub__(self, shift: int) -> "ChPoints":
+        """Subtracts the same scalar from all indices.
+
+        .. versionadded:: 0.0.3
+        """
+        return self + -shift
+
+    def plot(
+        self,
+        ax: _plt.Axes,
+        *,
+        t: _np.ndarray,
+        y: _np.ndarray,
+        points: _t.Literal[True] | set[str] = True,
+    ) -> _plt.Axes:
+        """Plots the points on the given axes ``ax`` using timestamps ``t`` and
+        amplitudes ``y``.
+
+        The purpose of this method is to be able to plot characteristic points
+        independently of the containing signal, useful for previewing results
+        of a detection algorithm.
+
+        For plotting the own characteristic points of a signal, please use
+        :meth:`Signal.plot`.
+
+        .. warning::
+            This method doesn't validate ``t`` and ``y`` against the indices.
+
+        :param ax: subplot axes.
+        :param t: timestamps.
+        :param y: amplitudes.
+        :param points: set to filter points by name. ``True`` means all.
+        :return: the axes.
+
+        .. versionadded:: 0.0.3
+        """
+        all_indices = _col.defaultdict(list)
+        for ci in self.indices:
+            for name, index in ci.without_unset().items():
+                if points is True or name in points:
+                    all_indices[name].append(index)
+        for name, indices in all_indices.items():
+            ax.plot(t[indices], y[indices], "+", label=name)
+        return ax
 
 
 class _MarksProxy(_col.UserDict):
@@ -141,14 +257,14 @@ class _SlicesProxy(_col.UserDict):
     def __setitem__(self, key: str, value: _ca.Sequence[slice]) -> None:
         for i, slc in enumerate(value):
             if slc.step is not None:
-                warnings.warn(f"Handling `step` is not implemented (value[{i}])")
+                warnings.warn(f"Handling `step` is not implemented (value[{key}][{i}])")
             if slc.start < 0 or slc.start >= self._stop:
                 raise ValueError(
-                    f"`value[{i}].start must be valid nonnegative index of `y`"
+                    f"`value[{key}][{i}].start must be valid nonnegative index of `y`"
                 )
             if slc.stop < 0 or slc.stop > self._stop:
                 raise ValueError(
-                    f"`value[{i}].stop must be valid nonnegative index of `y`"
+                    f"`value[{key}][{i}].stop must be valid nonnegative index of `y`"
                 )
         super().__setitem__(key, list(value))
 
@@ -474,42 +590,28 @@ class Signal:
         else:
             stop = slc.stop if slc.stop >= 0 else length + slc.stop
 
+        def crop_slices(slcs) -> list[slice]:
+            cropped = [
+                slice(max(0, s.start - start), min(s.stop - start, stop - start))
+                for s in slcs
+                if start <= s.start < stop or start <= s.stop <= stop
+            ]
+            # Remove empty slices
+            cropped = [s for s in cropped if s.stop > s.start]
+            return cropped
+
         section = type(self)(
             y=self.y[slc],
             unit=self.unit,
             t=self.t[slc],  # intentionally the t property
             fs=None,
             label=self.label,
-            chpoints=(
-                None
-                if self.chpoints is None
-                else _dc.replace(
-                    self.chpoints,
-                    indices=[
-                        ci.clamped(start, stop) - start
-                        for ci in self.chpoints.indices
-                        if (
-                            (ci.min() >= start and ci.max() < stop)
-                            or (ci.min() <= start <= ci.max())
-                            or (ci.min() < stop <= ci.max())
-                        )
-                    ],
-                )
-            ),
+            chpoints=(None if self.chpoints is None else self.chpoints[start:stop]),
             marks={
                 name: v[(v >= start) & (v < stop)] - start
                 for name, v in self.marks.items()
             },
-            slices={
-                name: [
-                    slice(
-                        max(0, slc.start - start), min(slc.stop - start, stop - start)
-                    )
-                    for slc in slcs
-                    if start <= slc.start < stop or start <= slc.stop <= stop
-                ]
-                for name, slcs in self.slices.items()
-            },
+            slices={name: crop_slices(slcs) for name, slcs in self.slices.items()},
             meta=self.meta,
         )
         section._fs = self._fs
@@ -666,6 +768,7 @@ class Signal:
         marks: bool | set[str] = True,
         legend: str = "auto",
         title: str | None = None,
+        t0: float | None = None,
         **plot_kw,
     ) -> _plt.Axes:
         """Plots the signal.
@@ -679,9 +782,16 @@ class Signal:
         :param legend: ``auto``: show legend when needed,
             ``off``: don't show, ``outside``: move legend out of plot area.
         :param title: plot title.
+        :param t0: shift timestamps to start from ``t0``, without copying or
+            modifying the object (unlike :meth:`shift_t`).
+
+            .. versionadded:: 0.0.3
         :param fmt_args: line style and color shorthand for :func:`visu.plot_signal`.
         :param plot_kw: keyword arguments for :func:`visu.plot_signal`.
         :return: the axes object.
+
+        .. versionchanged:: 0.0.3
+            Grid is on by default.
         """
         if not ax:
             _, ax = _plt.subplots()
@@ -693,7 +803,7 @@ class Signal:
 
         _v.plot_signal(
             ax,
-            self.t,
+            self.t if t0 is None else self.t - self.t[0] + t0,
             self.y,
             *fmt_args,
             y_unit=self.unit,
@@ -702,15 +812,13 @@ class Signal:
         )
 
         if points and self.chpoints:
-            all_indices = _col.defaultdict(list)
-            for ci in self.chpoints.indices:
-                for name, index in ci.without_unset().items():
-                    if (points is True or name in points) and not (
-                        onsets and name == "onset"
-                    ):
-                        all_indices[name].append(index)
-            for name, indices in all_indices.items():
-                ax.plot(self.t[indices], self.y[indices], "+", label=name)
+            points_: bool | set[str]
+            if onsets:
+                points_ = set(CpIndices.NAMES)
+                points_ = points_ - {"onset"}
+            else:
+                points_ = points
+            self.chpoints.plot(ax, t=self.t, y=self.y, points=points_)
             has_legend = True
         if onsets and self.chpoints and len(self.onsets):
             ax.plot(
@@ -749,6 +857,9 @@ class Signal:
                 ax.legend()
         if title:
             ax.set(title=title)
+
+        ax.grid(True)
+
         return ax
 
     def to_hdf(self, group: h5py.Group) -> None: